transferred topological sort into its own C file
