+++ src/graph_traversals/graph_traversals.h
@@ -17,6 +17,8 @@ void add_edge_undirected(Graph* graph, int src, int dest);
 void free_graph(Graph* graph);
 void print_graph(const Graph* graph);
 void add_edge_directed_unweighted(Graph* graph, int src, int dest);
+
+// ------------------For topological sort----------------------
 void topological_sort_kahn(Graph* graph);
 void topological_sort_demo(void);
